Silencing conversion warning.

+++ ch23/main.cpp
@@ -6,7 +6,7 @@ int main()
     MyVector<int> vector(10);
 
     for (std::size_t i{ }; i < 10; ++i)
-        vector[i] = i+1;
+        vector[i] = static_cast<int>(i+1);
 
     vector.resize(8);
     vector.insert(20, 5);
